try it better

+++ Aplicacion/Cursos/Eliminar.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
-using Aplicacion.ErrorHandling;
+using Aplicacion.ManejadorError;
 using MediatR;
 using Persistencia;
 
@@ -11,50 +11,50 @@ namespace Aplicacion.Cursos
 {
     public class Eliminar
     {
-        public class EliminarCursoRequest : IRequest
-        {
-            public Guid CursoId { get; set;}
+        public class Ejecuta : IRequest {
+            public Guid Id {get;set;}
         }
 
-        public class EliminarCursoRequestHandler : IRequestHandler<EliminarCursoRequest>
+        public class Manejador : IRequestHandler<Ejecuta>
         {
-            public readonly CursosOnlineContext _context;
-            public EliminarCursoRequestHandler(CursosOnlineContext context)
-            {
+            private readonly CursosOnlineContext _context;
+            public Manejador(CursosOnlineContext context){
                 _context = context;
             }
-
-            public async Task<Unit> Handle(EliminarCursoRequest request, CancellationToken cancellationToken)
+            public async Task<Unit> Handle(Ejecuta request, CancellationToken cancellationToken)
             {
-                var curso = await _context.Curso.FindAsync(request.CursoId) ?? throw new ExceptionHandling(HttpStatusCode.NotFound, new { message = "No se encontró el curso" });
-
-                // Remover instructores de la tabla cursoinstructor
-                var instructoresDB = _context.CursoInstructor.Where(x => x.CursoId == request.CursoId);
-                foreach (var instructor in instructoresDB)
-                {
+                var instructoresDB = _context.CursoInstructor.Where(x=> x.CursoId == request.Id);
+                foreach(var instructor in instructoresDB){
                     _context.CursoInstructor.Remove(instructor);
                 }
                 
-                // Remover precio
-                var precioDB = _context.Precio.Where(x => x.CursoId == request.CursoId).FirstOrDefault();
-                _context.Precio.Remove(precioDB);
-
-                // Remover comentarios
-                var comentariosDB = _context.Comentario.Where(x => x.CursoId == request.CursoId);
-                foreach (var comentario in comentariosDB)
-                {
-                    _context.Comentario.Remove(comentario);
+                var comentariosDB = _context.Comentario.Where(x=>x.CursoId == request.Id);
+                foreach(var cmt in comentariosDB){
+                    _context.Comentario.Remove(cmt);
                 }
 
-                _context.Curso.Remove(curso);
+                var precioDB = _context.Precio.Where(x=>x.CursoId == request.Id).FirstOrDefault();
+                if(precioDB!=null){
+                    _context.Precio.Remove(precioDB);
+                }
 
-                var valor = await _context.SaveChangesAsync();
-                
-                if (valor > 0)
-                {
+
+                var curso = await _context.Curso.FindAsync(request.Id);
+                if(curso==null){
+                    //throw new Exception("No se puede eliminar curso");
+                    throw new ManejadorExcepcion(HttpStatusCode.NotFound, new {mensaje = "No se encontro el curso"});
+                }
+                _context.Remove(curso);
+
+             
+
+                var resultado = await _context.SaveChangesAsync();
+
+                if(resultado>0){
                     return Unit.Value;
                 }
-                throw new Exception("No se pudo eliminar el curso");
+
+                throw new Exception("No se pudieron guardar los cambios");
             }
         }
     }
